refactor: add parent class of detectors and remove conversion warnings
add new base class to be the parent class of two-average and wqrs detectors
fix code to reduce the compiles warning caused by -Wconversion

+++ src/algorithms/mwa.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 
 
-std::vector<double> MWA_cumulative(const std::vector<double>& input_array, int window_size) {
+std::vector<double> MWA_cumulative(const std::vector<double>& input_array, size_t window_size) {
     std::vector<double> ret(input_array.size(), 0.0);
     std::partial_sum(input_array.begin(), input_array.end(), ret.begin());
 
@@ -11,17 +11,17 @@ std::vector<double> MWA_cumulative(const std::vector<double>& input_array, int w
         ret[i] = ret[i] - ret[i - window_size];
     }
 
-    for (int i = 1; i < window_size; ++i) {
-        ret[i - 1] = ret[i - 1] / i;
+    for (size_t i = 1; i < window_size; ++i) {
+        ret[i - 1] = ret[i - 1] / static_cast<double>(i);
     }
     for (size_t i = window_size - 1; i < ret.size(); ++i) {
-        ret[i] = ret[i] / window_size;
+        ret[i] = ret[i] / static_cast<double>(window_size);
     }
 
     return ret;
 }
 
-std::vector<double> MWA_original(const std::vector<double>& input_array, int window_size) {
+std::vector<double> MWA_original(const std::vector<double>& input_array, size_t window_size) {
     std::vector<double> mwa(input_array.size(), 0.0);
     double sum = 0.0;
 
@@ -31,16 +31,16 @@ std::vector<double> MWA_original(const std::vector<double>& input_array, int win
             sum -= input_array[i - window_size];
         }
         if (i < window_size - 1) {
-            mwa[i] = sum / (i + 1);
+            mwa[i] = sum / static_cast<double>(i + 1);
         } else {
-            mwa[i] = sum / window_size;
+            mwa[i] = sum / static_cast<double>(window_size);
         }
     }
 
     return mwa;
 }
 
-std::vector<double> MWA_convolve(const std::vector<double>& input_array, int window_size) {
+std::vector<double> MWA_convolve(const std::vector<double>& input_array, size_t window_size) {
     std::vector<double> padded_input(window_size - 1, 0.0);
     padded_input.insert(padded_input.end(), input_array.begin(), input_array.end());
 
@@ -51,11 +51,11 @@ std::vector<double> MWA_convolve(const std::vector<double>& input_array, int win
         ret[i] = std::inner_product(padded_input.begin() + i, padded_input.begin() + i + window_size, window.begin(), 0.0);
     }
 
-    for (int i = 1; i < window_size; ++i) {
-        ret[i - 1] = ret[i - 1] / i;
+    for (size_t i = 1; i < window_size; ++i) {
+        ret[i - 1] = ret[i - 1] / static_cast<double>(i);
     }
     for (size_t i = window_size - 1; i < ret.size(); ++i) {
-        ret[i] = ret[i] / window_size;
+        ret[i] = ret[i] / static_cast<double>(window_size);
     }
 
     return ret;
